[gaeun] 출항일이 지난 것을 3일 이후로 변경

+++ src/berth-py/berth-py.service.ts
@@ -62,28 +62,30 @@ export class BerthPyService {
     return await this.seqeulize.query(
       `
       SELECT 
-       oid
+        oid
       FROM
-      berthStat_schedule
+          berthStat_schedule
       WHERE
-        IF(
-          LEFT(tkoffPrarnde, 1
-          )
-        = '(',
-        MID(
-          tkoffPrarnde, 2, 16
-        ),
-        LEFT(
-          tkoffPrarnde, 19)
-        )
-        BETWEEN (
-          SELECT 
-            DATE_FORMAT(DATE_ADD(NOW(), INTERVAL - 7 DAY), '%Y-%m-%d')
-          FROM DUAL)
-        AND
-        (SELECT DATE_FORMAT(NOW(), '%Y-%m-%d') FROM DUAL)
+        TRUE
+        -- 출항일이 3일 지난 것만
+        AND DATE_FORMAT(IF(LEFT(tkoffPrarnde, 1) = '(',
+                MID(tkoffPrarnde, 2, 16),
+                LEFT(tkoffPrarnde, 19)),
+            '%Y-%m-%d %H:%i') IN (IF(DATE_ADD(DATE_FORMAT(IF(LEFT(tkoffPrarnde, 1) = '(',
+                    MID(tkoffPrarnde, 2, 16),
+                    LEFT(tkoffPrarnde, 19)),
+                '%Y-%m-%d %H:%i'), INTERVAL 3 DAY) < DATE_FORMAT(NOW(), '%Y-%m-%d %H:%i'),
+        DATE_FORMAT(IF(LEFT(tkoffPrarnde, 1) = '(',
+                    MID(tkoffPrarnde, 2, 16),
+                    LEFT(tkoffPrarnde, 19)),
+                '%Y-%m-%d %H:%i'),
+        NULL))
       `,
-      { type: sequelize.QueryTypes.SELECT }
+      {
+        type: sequelize.QueryTypes.SELECT,
+        model: berthStatSchedule,
+        mapToModel: true,
+      }
     );
   }
 
@@ -177,12 +179,12 @@ export class BerthPyService {
               { where: { oid: obj.oid }, transaction: t }
             );
 
-            // /** 입항일자 변경으로 인한 문자 전송 - 테스트 기간까지는 문자 전송을 하지 않음 */
-            // await this.sendAlramOfcsdhpPrarnde(
-            //   userInfoList,
-            //   obj,
-            //   berthDupleData
-            // );
+            /** 입항일자 변경으로 인한 문자 전송 - 테스트 기간까지는 문자 전송을 하지 않음 */
+            await this.sendAlramOfcsdhpPrarnde(
+              userInfoList,
+              obj,
+              berthDupleData
+            );
 
             /** 알람 메시지 create */
             await this.sendWebAlramOfcsdhpPrarnde(
@@ -211,9 +213,14 @@ export class BerthPyService {
       /** 삭제할 선석 데이터 */
       const getAllBerthOldDataList = await this.findAllBerthOldDataList();
 
+      if (getAllBerthOldDataList.length === 0) {
+        await t.rollback();
+        return;
+      }
+
       for (const obj of getAllBerthOldDataList) {
         await berthStatSchedule.destroy({
-          where: { oid: obj },
+          where: { oid: obj.oid },
           transaction: t,
         });
       }
